requirements.txt
smtplib
email
tkinter
schedule==1.2.0
jinja2==3.1.2
pandas==2.0.3
python-dotenv==1.0.0